+++ src/main/kotlin/dev/jbang/idea/completion/directive/JBangJavaCompletionContributor.kt
@@ -1,19 +1,9 @@
 package dev.jbang.idea.completion.directive
 
-import com.intellij.codeInsight.completion.CompletionParameters
-import com.intellij.codeInsight.completion.CompletionProvider
-import com.intellij.codeInsight.completion.CompletionResultSet
-import com.intellij.codeInsight.completion.CompletionType
-import com.intellij.codeInsight.lookup.LookupElementBuilder
 import com.intellij.lang.java.JavaLanguage
-import com.intellij.patterns.PlatformPatterns
-import com.intellij.psi.PsiClass
-import com.intellij.psi.PsiComment
-import com.intellij.psi.PsiFile
-import com.intellij.util.ProcessingContext
 
 
-class JBangJavaCompletionContributor : JBangBaseDirectiveCompletionContributor() {
+class JBangJavaCompletionContributor : JBangBaseDirectiveCompletionContributor(JavaLanguage.INSTANCE) {
 
 
 }