+++ src/main/kotlin/dev/jbang/idea/completion/directive/JBangKotlinCompletionContributor.kt
@@ -1,23 +1,18 @@
 package dev.jbang.idea.completion.directive
 
 import com.intellij.codeInsight.completion.CompletionParameters
-import com.intellij.codeInsight.completion.CompletionProvider
 import com.intellij.codeInsight.completion.CompletionResultSet
-import com.intellij.codeInsight.completion.CompletionType
 import com.intellij.codeInsight.lookup.LookupElementBuilder
-import com.intellij.patterns.PlatformPatterns
-import com.intellij.psi.PsiComment
-import com.intellij.psi.PsiFile
 import com.intellij.util.ProcessingContext
 import org.jetbrains.kotlin.idea.KotlinLanguage
-import org.jetbrains.kotlin.psi.KtClass
-import org.jetbrains.kotlin.psi.KtNamedFunction
 
-class JBangKotlinCompletionContributor : JBangBaseDirectiveCompletionContributor() {
-    override fun _addCompletions(parameters: CompletionParameters,
-                                context: ProcessingContext,
-                                result: CompletionResultSet) {
-        super._addCompletions(parameters, context, result)
+class JBangKotlinCompletionContributor : JBangBaseDirectiveCompletionContributor(KotlinLanguage.INSTANCE) {
+    override fun addCompletions(
+        parameters: CompletionParameters,
+        context: ProcessingContext,
+        result: CompletionResultSet
+    ) {
+        super.addCompletions(parameters, context, result)
         result.addElement(LookupElementBuilder.create("KOTLIN").withTailText("Kotlin version", true))
     }