+++ src/main/kotlin/dev/jbang/idea/completion/sources/JBangGroovySourcesCompletionContributor.kt
@@ -1,0 +1,18 @@
+package dev.jbang.idea.completion.sources
+
+import dev.jbang.idea.groovyIcon
+import org.jetbrains.plugins.groovy.GroovyLanguage
+import javax.swing.Icon
+
+
+class JBangGroovySourcesCompletionContributor : JBangBaseSourcesCompletionContributor(GroovyLanguage) {
+
+    override fun allowedExtName(): String {
+        return ".groovy"
+    }
+
+
+    override fun fileTypeIcon(): Icon {
+        return groovyIcon
+    }
+}